requirements.txt
python-fasthtml
fastcore